Add files via upload
This is an update to the first version. The factory pattern now works with the decorator pattern and the decorator classes have been fully labeled. I removed 2 of the interfaces and just made all of the items work from 1 interface. Using 3 interfaces for the items proved to be very cumbersome. I've created a demo of the factory creating the food items in the file TestFactory.java. It works with the main courses, but there are still some bugs.