Add files via upload changelog in description
Changelog:

Added boolean hasCondiment to determine whether to ask for condiments
Fixed a typo that caused the "Chips" item to break.
Fixed a typo that caused the "Soda" item to break.
Fixed a crash that would happen if "Unsweetened Tea" was selected.
Converted mcList, snList, and drList into Lists. They used to be Queues. This change will make it easier to edit the elements of those lists.